patch 7

+++ PitMystics/src/main/java/dev/kyro/pitsim/enchants/SpeedyHit.java
@@ -28,7 +28,7 @@ public class SpeedyHit extends PitEnchant {
 		Cooldown cooldown = getCooldown(attackEvent.attacker,(getCooldown(enchantLvl) * 20));
 		if(cooldown.isOnCooldown()) return; else cooldown.reset();
 
-		Misc.applyPotionEffect(attackEvent.attacker, PotionEffectType.SPEED, (int) (getDuration(enchantLvl) * 20), 0, true, false);
+		Misc.applyPotionEffect(attackEvent.attacker, PotionEffectType.SPEED, getDuration(enchantLvl) * 20, 0, true, false);
 	}
 
 	@Override
